presse

+++ src/CmsBundle/Controller/PresseController.php
@@ -23,6 +23,8 @@ class PresseController extends Controller
     {
         $em = $this->getDoctrine()->getManager();
         $id_item_max = $em->getRepository('CmsBundle:Presse')->getIdItemPresse();
+        $langue_active = $em->getRepository('CmsBundle:Accueil')->findBy(array('langue' => 'fr'))[0]->getLangueActive();
+
 
         $langue = new Presse();
 
@@ -31,9 +33,15 @@ class PresseController extends Controller
         $presse_fr = new Presse();
         $presse_fr->setLangue('fr');
         $langue->getPresse()->add($presse_fr);
-        $presse_en = new Presse();
-        $presse_en->setLangue('en');
-        $langue->getPresse()->add($presse_en);
+
+
+        if ($langue_active == true){
+            if (empty($accueil_en)) {
+                $presse_en = new Presse();
+                $presse_en->setLangue('en');
+                $langue->getPresse()->add($presse_en);
+            }
+        }
         // end dummy code
 
         $form = $this->createFormBuilder($langue)
@@ -47,16 +55,21 @@ class PresseController extends Controller
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             $presse_fr->setItemId($id_item_max[0][1] + 1);
+
+            if ($langue_active == true){
             $presse_en->setItemId($presse_fr->getItemId());
             $presse_en->setImage($presse_fr->getImage());
             $presse_en->setDate($presse_fr->getDate());
             $presse_en->setLien($presse_fr->getLien());
+            }
 
             $em->persist($presse_fr);
 
-            $presse_en->setImage($presse_fr->getImage());
+            if ($langue_active == true) {
+                $presse_en->setImage($presse_fr->getImage());
+                $em->persist($presse_en);
+            }
 
-            $em->persist($presse_en);
             $em->flush();
 
             return $this->redirectToRoute('user_presse');
@@ -64,6 +77,7 @@ class PresseController extends Controller
 
         return $this->render('@Cms/presse/new.html.twig', array(
             'form' => $form->createView(),
+            'langue_active' => $langue_active
         ));
     }
 
@@ -79,13 +93,22 @@ class PresseController extends Controller
 
         $presse_fr = $em->getRepository('CmsBundle:Presse')->findOneBy(array('langue' => 'fr', 'item_id' => $id_item));
         $presse_en = $em->getRepository('CmsBundle:Presse')->findOneBy(array('langue' => 'en', 'item_id' => $id_item));
+        $langue_active = $em->getRepository('CmsBundle:Accueil')->findBy(array('langue' => 'fr'))[0]->getLangueActive();
 
         $langue = new Presse();
 
         // dummy code - this is here just so that the Task has some tags
         // otherwise, this isn't an interesting example
         $langue->getPresse()->add($presse_fr);
-        $langue->getPresse()->add($presse_en);
+
+        if ($langue_active == true){
+            if (empty($presse_en) == false){
+                $presse_en = new Presse();
+                $presse_en->setLangue('en');
+                $presse_en->setItemId($presse_en->getItemId());
+            }
+            $langue->getPresse()->add($presse_en);
+        }
 
         $editForm = $this->createFormBuilder($langue)
             ->add('presse', CollectionType::class, array(
@@ -103,15 +126,19 @@ class PresseController extends Controller
             $em->persist($presse_fr);
             $em->flush();
 
-            $presse_en->setImage($presse_fr->getImage());
-            $em->persist($presse_en);
-            $em->flush();
+            if ($langue_active == true){
+                $presse_en->setDate($presse_fr->getDate());
+                $presse_en->setImage($presse_fr->getImage());
+                $em->persist($presse_en);
+                $em->flush();
+            }
 
             return $this->redirectToRoute('user_presse', array('id' => $presse->getId()));
         }
 
         return $this->render('CmsBundle:presse:edit.html.twig', array(
             'form' => $editForm->createView(),
+            'langue_active' => $langue_active
         ));
     }
     /**
@@ -134,7 +161,9 @@ class PresseController extends Controller
         }
 
         $em->remove($presse_fr);
-        $em->remove($presse_en);
+
+        if ($presse_en != null)
+            $em->remove($presse_en);
 
         $em->flush();
